+++ src/shared/SharedConfig.h
@@ -312,9 +312,6 @@ struct __attribute__((packed, aligned(8))) SharedConfig {
     /* Tries to detect busy loops and advance time */
     bool busyloop_detection = false;
 
-    /* User can modify the framerate during the game execution */
-    bool variable_framerate = false;
-
     /* Send stack traces of all time calls to libTAS program */
     bool time_trace = false;